+++ modules/web/containers/root.js
@@ -1,6 +1,6 @@
 import React, { PropTypes, cloneElement } from 'react'
 import { Provider } from 'react-redux'
-import Notifications from './notifications'
+import Notifications from '../modules/notifications/notifications'
 
 const Root = props => (
 	<Provider store={props.store}>
