+++ modules/web/modules/notifications/notification.scss
@@ -1,5 +1,5 @@
-@import '../styles/variables.scss';
-@import '../styles/mixins.scss';
+@import '../../styles/variables';
+@import '../../styles/mixins';
 
 .notification-capsule {
 	position: relative;
