+++ modules/web/modules/notifications/notifications.js
@@ -3,9 +3,9 @@ import React, { PropTypes } from 'react'
 import map from 'lodash/map'
 import { bindActionCreators } from 'redux'
 import { connect } from 'react-redux'
-import { removeNotification } from '../redux/notifications/actions'
+import { removeNotification } from './redux/actions'
 
-import Notification from '../components/notification'
+import Notification from './notification'
 
 import './notifications.scss'
