+++ modules/web/modules/notifications/notifications.scss
@@ -1,4 +1,4 @@
-@import '../styles/variables.scss';
+@import '../../styles/variables';
 
 .notification-enter {
 	opacity: 0.01;
