+++ modules/web/modules/notifications/redux/__tests__/.eslintrc.yml
@@ -1,0 +1,2 @@
+---
+extends: ../../../../../../.eslintrc-test.yml
