--- modules/web/redux/notifications/__tests__/.eslintrc.yml
@@ -1,2 +1,0 @@
----
-extends: ../../../../../.eslintrc-test.yml
