+++ modules/web/redux/reducer.js
@@ -2,7 +2,7 @@
 import { combineReducers } from 'redux'
 
 import areas from './areas/reducers'
-import notifications from './notifications/reducers'
+import notifications from '../modules/notifications/redux/reducers'
 import search from '../modules/course-searcher/redux/reducers'
 import students from './students/reducers'
